Implemented validation for image using the required attribute. closes #112
