+++ BrickPile.Domain/BrickPile.Domain.csproj
@@ -71,12 +71,9 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Models\Asset.cs" />
-    <Compile Include="Models\Audio.cs" />
-    <Compile Include="Models\Document.cs" />
     <Compile Include="Models\DocumentReference.cs" />
     <Compile Include="Models\IContent.cs" />
     <Compile Include="Models\IDocumentReference.cs" />
-    <Compile Include="Models\Image.cs" />
     <Compile Include="Models\IPageMetadata.cs" />
     <Compile Include="Models\IPageModel.cs" />
     <Compile Include="Models\PageMetadata.cs" />
@@ -88,7 +85,6 @@
     <Compile Include="Models\Thumbnail.cs" />
     <Compile Include="Models\Thumbnails.cs" />
     <Compile Include="Models\User.cs" />
-    <Compile Include="Models\Video.cs" />
     <Compile Include="ContentTypeAttribute.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ResourceHelper.cs" />
