+++ BrickPile.UI/Areas/UI/Views/Shared/EditorTemplates/Image.cshtml
@@ -1,7 +1,7 @@
 ﻿@using System.Configuration
 @using System.Web.Configuration
 @using BrickPile.UI.Common
-@model BrickPile.Domain.Models.Image
+@model BrickPile.UI.Web.Image
 
 @{
     var maxRequestLength = 4194304;
