+++ BrickPile.UI/BrickPile.UI.csproj
@@ -165,10 +165,13 @@
     <Compile Include="IStructureInfo.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="StructureInfo.cs" />
+    <Compile Include="Web\Audio.cs" />
     <Compile Include="Web\ContainerDisposalModule.cs" />
+    <Compile Include="Web\Document.cs" />
     <Compile Include="Web\Hosting\NativeVirtualDirectory.cs" />
     <Compile Include="Web\Hosting\NativeVirtualFile.cs" />
     <Compile Include="Web\Hosting\NativeVirtualPathProvider.cs" />
+    <Compile Include="Web\Image.cs" />
     <Compile Include="Web\Mvc\BrickPileControllerFactory.cs" />
     <Compile Include="Web\Mvc\ContentModelBinder.cs" />
     <Compile Include="Web\Mvc\ControllerMapper.cs" />
@@ -192,6 +195,7 @@
     <Compile Include="Web\Security\PasswordUtil.cs" />
     <Compile Include="Web\Security\RavenDBMembershipProvider.cs" />
     <Compile Include="Web\Security\RavenDBRoleProvider.cs" />
+    <Compile Include="Web\Video.cs" />
     <Compile Include="Web\ViewModels\DefaultViewModel.cs" />
     <Compile Include="Web\ViewModels\EditViewModel.cs" />
     <Compile Include="Web\ViewModels\IndexViewModel.cs" />
