+++ BrickPile.UI/Global.asax.cs
@@ -1,6 +1,8 @@
-﻿using System.Web.Http;
+﻿using System.Web.Hosting;
+using System.Web.Http;
 using System.Web.Mvc;
 using System.Web.Routing;
+using BrickPile.UI.Web.Hosting;
 
 namespace BrickPile.Samples {
     // Note: For instructions on enabling IIS6 or IIS7 classic mode, 
@@ -12,6 +14,8 @@ namespace BrickPile.Samples {
             WebApiConfig.Register(GlobalConfiguration.Configuration);
             FilterConfig.RegisterGlobalFilters(GlobalFilters.Filters);
             RouteConfig.RegisterRoutes(RouteTable.Routes);
+
+            HostingEnvironment.RegisterVirtualPathProvider(new NativeVirtualPathProvider());
 
         }
     }
