+++ BrickPile.UI/Web/Audio.cs
@@ -1,4 +1,6 @@
-namespace BrickPile.Domain.Models {
+using BrickPile.Domain.Models;
+
+namespace BrickPile.UI.Web {
     public class Audio : Asset {
         public Audio() {
             Id = "assets/audios/";
