+++ BrickPile.UI/Web/Document.cs
@@ -1,4 +1,6 @@
-namespace BrickPile.Domain.Models {
+using BrickPile.Domain.Models;
+
+namespace BrickPile.UI.Web {
     public class Document : Asset {
         /// <summary>
         /// Initializes a new instance of the <see cref="Document"/> class.
