+++ BrickPile.UI/Web/Video.cs
@@ -1,4 +1,6 @@
-namespace BrickPile.Domain.Models {
+using BrickPile.Domain.Models;
+
+namespace BrickPile.UI.Web {
     public class Video : Asset {
         /// <summary>
         /// Initializes a new instance of the <see cref="Video"/> class.
